Fix warnings in test code

+++ src/fast-aleck/test/fast-aleck-test.c
@@ -697,8 +697,8 @@ static void fast_aleck_test(struct fa_test_suite *a_test_suite, char *a_input, c
 		++a_test_suite->fails;
 		fprintf(stdout, "not ok %i ", a_test_suite->count+1);
 		_fa_puts_escaped(a_input);
-		fprintf(stdout, "  Length of returned string: %lu\n", strlen(actual_output));
-		fprintf(stdout, "  Returned length of string: %lu\n", out_len);
+		fprintf(stdout, "  Length of returned string: %lu\n", (unsigned long) strlen(actual_output));
+		fprintf(stdout, "  Returned length of string: %lu\n", (unsigned long) out_len);
 		goto bail;
 	}
 
@@ -731,7 +731,7 @@ static void fast_aleck_test(struct fa_test_suite *a_test_suite, char *a_input, c
 		++a_test_suite->fails;
 		fprintf(stdout, "not ok %i ", a_test_suite->count+1);
 		_fa_puts_escaped(a_input);
-		fprintf(stdout, "  Expected number of tokens: %lu\n", expected_count);
+		fprintf(stdout, "  Expected number of tokens: %lu\n", (unsigned long) expected_count);
 		fprintf(stdout, "  Expected tokens:\n");
 		for (size_t i = 0; i < expected_count; ++i) {
 			fa_token expected_token = fa_token_buffer_at(&expected_token_buffer, i);
@@ -740,7 +740,7 @@ static void fast_aleck_test(struct fa_test_suite *a_test_suite, char *a_input, c
 			fprintf(stdout, " ");
 			puts("");
 		}
-		fprintf(stdout, "  Actual number of tokens:   %lu\n", actual_count);
+		fprintf(stdout, "  Actual number of tokens:   %lu\n", (unsigned long) actual_count);
 		fprintf(stdout, "  Actual tokens:  \n");
 		for (size_t i = 0; i < actual_count; ++i) {
 			fa_token actual_token = fa_token_buffer_at(&actual_token_buffer, i);
@@ -759,10 +759,10 @@ static void fast_aleck_test(struct fa_test_suite *a_test_suite, char *a_input, c
 			++a_test_suite->fails;
 			fprintf(stdout, "not ok %i ", a_test_suite->count+1);
 			_fa_puts_escaped(a_input);
-			fprintf(stdout, "  Expected token %lu: ", i);
+			fprintf(stdout, "  Expected token %lu: ", (unsigned long) i);
 			fa_token_print(expected_token, stdout);
 			puts("");
-			fprintf(stdout, "  Actual token %lu:   ", i);
+			fprintf(stdout, "  Actual token %lu:   ", (unsigned long) i);
 			fa_token_print(actual_token, stdout);
 			puts("");
 			goto bail;
